Core climate model integration (#109)
* Added initial implementation of climate prediction model

* data type bugfixes (intellisense not working for ClimatePredictionModel file)

* recreate slnx files

* minor change

* add ml and physics model, with climate model interface class

* Add Game manager for AI

* temp save before pull from other remote branch

* completed callable climate model entrypoint. updated physics model to predict either temp, sea level, or both. made more progress on the ML model using onnxruntime (still incomplete). integrated GameManager wiht the climate model, with functionality to call the entry method to the climate model prediction method.

* minor changes

* implemented model prediction using onnxruntime. climate module bugfixes

* update tensor comment

* remove climate data class

* fix memory leak bug by disposing the inference session object on MLModel object dispose

* removed unused tensor output name string in MLModel.cs

* add and fix comments in mlmodel

* Add logic to get the UI to use to OnChange event for the vlaue

* Apply changes to PlayerManager prefab

* Change SEA_LEVEL_SCALE_FACTOR for metres to and from millimetres

* Fixed some Bugs

Adjusted pollution level of buildings made work with model :)))))))))))))

* Model changes and added debugs for using ml or physics model


* move setting of climate related network obj variables out of awake function and set them to initial values on spawn

---------

+++ Assets/Scripts/DataClasses/TileFeatures/FeatureType.cs
@@ -39,7 +39,7 @@ namespace TTT.DataClasses.TileFeatures
         /// </summary>
         [field: Tooltip("The amount of pollution emitted per turn.")]
         [field: SerializeField]
-        public int PollutionEmission { get; private set; }
+        public float PollutionEmission { get; private set; }
 
         /// <summary>
         /// The constraints for building this feature.
